move toast component to common folder
