+++ bigbluebutton-html5/imports/ui/components/status-notifier/styles.js
@@ -20,7 +20,7 @@ import {
   fontSizeSmall,
 } from '/imports/ui/stylesheets/styled-components/typography';
 import Button from '/imports/ui/components/common/button/component';
-import ToastStyled from '/imports/ui/components/toast/styles';
+import ToastStyled from '/imports/ui/components/common/toast/styles';
 
 const Avatar = styled.div`
   cursor: pointer;
